+++ src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import { BrowserRouter as Router, Switch, Route, Link, useParams, useRouteMatch } from 'react-router-dom';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import NavBar from './components/navbar/NavBar';
 import Footer from './components/footer/Footer';
 import HomePage from './components/home/HomePage';
