+++ src/components/login-signup/LoginPage.js
@@ -1,6 +1,7 @@
-import React, { useState }  from 'react';
+import React, { useEffect, useState }  from 'react';
 import {Redirect} from 'react-router-dom';
 import {ReactSession as Session} from 'react-client-session';
+import {getEntry, getMethod, table} from '../../helpers/AxiosHelper';
 
 export default function LoginPage(props){
     //Checks to see if user can access page
@@ -9,55 +10,80 @@ export default function LoginPage(props){
     //State constants
     const [username, setUsername] = useState("");
     const [password, setPassword] = useState("");
+    const [tableChoice, setTableChoice] = useState("");
     const [error, setError] = useState("");
-    const handleSubmit = (event) => {
-        callAPI(username, password, setError);
+    const [redirect, setRedirect] = useState("");
+    const [loading, setLoading] = useState("");
+    const [resultData, setResultData] = useState([]);
+
+    //this polls the database for the username given when a button is hit
+    useEffect(() => {
+        if (tableChoice !== "") {
+            const fetchData = async function() {
+                try {
+                    setLoading(true);
+                    // console.log(window.btoa("anorthouse:password"));
+                    // console.log(url);
+                    const response = await getEntry(tableChoice, getMethod.BY_USERNAME, username);
+                    if (response.status === 200) {
+                        setResultData(response);
+                    }
+                } catch (error) {
+                    setResultData(false);
+                    throw error;
+                } finally {
+                    setLoading(false);
+                }
+            };
+            fetchData();
+        }
+    }, [tableChoice]);
+
+    //this then takes that polled information and goes from there.
+    useEffect(() => {
+        if (tableChoice !== "") {
+            //check for bad username
+            if (resultData.username === "N/A") {
+                console.log("invalid username");
+            }
+    
+            //check password
+    
+            //send redirect
+    
+            //{ userId: -1, username: "N/A", password: "N/A", name: "Anonymous", reviewCount: 0, role: "USER" }
+        }
+    }, [resultData]);
+
+    //button events
+    const userLogin = (event) => {
+        setTableChoice(table.USER);
         event.preventDefault();
     }
-
-    const changeUsername = (event) => {
-        setUsername(event.target.value);
+    const adminLogin = (event) => {
+        setTableChoice(table.ADMIN);
+        event.preventDefault();
     }
-    const changePassword = (event) => {
-        setPassword(event.target.value);
-    }
-
-    let errors = error === "" ?
-        <span></span> :
-        <p className="error">{error}</p> ;
 
     return needRedirect ?
         (<Redirect to="/"/>) :
         (<div>
-            <form onSubmit={handleSubmit}>
+            <form>
                 <label htmlFor="username-input">Username:</label>
                     <input type="text" id="username-input" name="username-input"
-                            value={username} onChange={changeUsername}></input>
+                            value={username} onChange={event => setUsername(event.target.value)}></input>
 
                 <label htmlFor="password-input">Password:</label>
                     <input type="text" id="password-input" name="password-input"
-                            value={password} onChange={changePassword}></input>
+                            value={password} onChange={event => setPassword(event.target.value)}></input>
 
-                <button type="submit" value="submit">Submit</button>
+                <button type="submit" onClick={userLogin}>User Login</button>
+                <button type="submit" onClick={adminLogin}>Admin Login</button>
             </form>
-            {errors}
+            <div>{resultData}</div>
+            <div>{redirect}</div>
+            <div>{loading}</div>
+            <div>{error}</div>
         </div>
     );
-}
-
-
-async function callAPI(username, password, setError) {
-    try {
-        console.log(username);
-        console.log(password);
-        const res = await fetch("http://swapi.dev/api/people/1");
-        if (!res.ok) {
-            throw Error(res.statusText);
-        }
-        const json = await res.json();
-        setError(json.name);
-    } catch (err) {
-        console.log(err);
-        setError(err.message);
-    }
 }
