+++ src/components/login-signup/LogoutPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import {Redirect} from 'react-router-dom';
-import {ReactSession as Session} from 'react-client-session';
+// import {Redirect} from 'react-router-dom';
+// import {ReactSession as Session} from 'react-client-session';
 
 export default function LogoutPage(props){
