+++ src/components/navbar/conditionals/LoginSignUp.js
@@ -1,4 +1,4 @@
-import React, { useState }  from 'react';
+import React  from 'react';
 import { Link } from 'react-router-dom';
 import {ReactSession as Session} from 'react-client-session';
